+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 ----- 0.12.0
 * Downgrade SDK to 23 because the NDK is buggy in 24
 ----- 0.12.1
+* fix keyboard handling with emoji chooser